\name{birth}
\docType{data}
\alias{birth}
\usage{data(birth)}
\title{Birth data}
\description{
Birth data for singleton live births with gestational age at least 38 weeks.
}
\format{A data frame with 400 observations on 9 variables.
  \tabular{lll}{
    \code{Sex}   \tab character  \tab 'male' or 'female' \cr
    \code{Gestation}  \tab numeric  \tab Gestational age (in weeks). \cr
    \code{Weight}   \tab numeric  \tab birth weight. \cr
    \code{Length}   \tab numeric  \tab height. \cr
    \code{Head}   \tab numeric  \tab head size. \cr
    \code{Chest}   \tab numeric  \tab chest size. \cr
    \code{Mother.s.age}   \tab numeric  \tab chest size. \cr
    \code{type}   \tab factor  \tab 'r' = rural or 'u' = urban. \cr
    \code{region}   \tab factor  \tab region of the birth. \cr
  }
}
\references{
  
  Wang, B. and Wertelecki, W. (2012) Density Estimation for Data With
  Rounding Errors, Computational Statistics and Data Analysis, doi:
  10.1016/j.csda.2012.02.016.

}
\keyword{datasets}
