% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainWrappers.R
\name{computeCorPosteriorLine}
\alias{computeCorPosteriorLine}
\title{Title}
\usage{
computeCorPosteriorLine(
  bfObject,
  alternative = "two.sided",
  minX = -0.985,
  maxX = 0.985
)
}
\arguments{
\item{maxX}{}
}
\description{
Title
}
