% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainWrappers.R
\name{computeCorRobustnessLine}
\alias{computeCorRobustnessLine}
\title{Title}
\usage{
computeCorRobustnessLine(bfObject)
}
\arguments{
\item{method}{}
}
\description{
Title
}
\examples{
bfObject <- bcor.testSumStat(n=34, stat=0.3)
result <- computeCorRobustnessLine(bfObject)

xLine <- result$kappaDomain
yLine <- result$two.sided$robustnessLine
yMax <- result$two.sided$robustnessMaxBf
plot(xLine, yLine, ylim=c(0, yMax), type="l")
}
