% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainWrappers.R
\name{computeCorSequentialLine}
\alias{computeCorSequentialLine}
\title{Title}
\usage{
computeCorSequentialLine(x, y, bfObject)
}
\arguments{
\item{method}{}
}
\description{
Title
}
\examples{
x <- rnorm(100)
y <- rnorm(100)
bfObject <- bcor.test(x, y, method="kendall")

computeCorSequentialLine <- computeCorSequentialLine(x, y, bfObject)
graphics::plot(computeCorSequentialLine$nDomain,
               log(computeCorSequentialLine$less$sequentialLine),
               type="l", xlab="n", ylab=expression(log("BF"[10])))
}
