\name{edu75}
\docType{data}
\alias{edu75}
\usage{data(edu75)}
\title{Education expenditure data (1975)}
\description{
Education expenditure data for all 50 states in U.S.A in 1975.}
\format{A data frame with 50 observations on 6 variables.
  \tabular{lll}{
    \code{States}   \tab character  \tab Initial of state names \cr
    \code{Y}  \tab numeric  \tab Educational expenditure. \cr
    \code{X1}   \tab numeric  \tab X1. \cr
    \code{X2}   \tab numeric  \tab X2. \cr
    \code{X3}   \tab numeric  \tab X3. \cr
    \code{Region}   \tab character  \tab region, 1=northwest, 2,3,4. \cr
  }
}
\references{
Stat 335 text
}
\keyword{datasets}
