% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainWrappers.R
\name{getSidedObject}
\alias{getSidedObject}
\title{Function to grab sided ("two.sided", "greater", "less") information of an bfObject with extra items}
\usage{
getSidedObject(bfObject, alternative = "two.sided", itemNames = NULL)
}
\arguments{
\item{itemNames}{optional character strings, that specify additional items from bfObject}
}
\value{
a (sub)list of bfObject
}
\description{
Function to grab sided ("two.sided", "greater", "less") information of an bfObject with extra items
}
