% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0helpers.R
\name{isSomeInfinite}
\alias{isSomeInfinite}
\title{Checks whether some object is infinite}
\usage{
isSomeInfinite(obj)
}
\value{
Returns TRUE if there's some object that's infinite, FALSE when all objects are finite
}
\description{
Checks whether some object is infinite
}
\examples{

isSomeInfinite(1:10)
isSomeInfinite(c(1e2000, 1:1000000))

# Faster than
# isSomeInfinite(c(1:1000000, 1e2000))
}
