% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson.R
\name{marsmanCorMHSampler}
\alias{marsmanCorMHSampler}
\title{Fits a beta density to the the posterior samples based on an independent Metropolis sampler, see Tierney (1994).}
\usage{
marsmanCorMHSampler(n, r, kappa = 1, nIters = 50000L)
}
\arguments{
\item{kappa}{numeric > 0 that specifies the}
}
\value{
Returns a list with the beta parameters of a stretched beta distribution on (-1, 1) and the acceptance rate.
}
\description{
Fits a beta density to the the posterior samples based on an independent Metropolis sampler, see Tierney (1994).
}
