% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson.R
\name{mPlusCorJeffreysIntegrate}
\alias{mPlusCorJeffreysIntegrate}
\title{Add this to the semi-exact two-sided Bayes factor to get bfPlus0}
\usage{
mPlusCorJeffreysIntegrate(n, r, kappa = 1, ...)
}
\arguments{
\item{n}{integer representing sample size}

\item{r}{numeric in (-1, 1) sample Pearson correlation r}

\item{kappa}{numeric > 0 sample}
}
\value{
Returns bf10, bfPlus0, bfMin0 and the stretched beta fit of the posterior.

Exact BF10 for Pearson's correlation based on the exact reduced likelihood, see Ly et al. (2018).

Returns Bayes factor BF10 in favour of the alternative over the null
}
\description{
#' Exact bf10, bfPlus0, bfMin0 based on the exact
}
\details{
for Pearson's correlation based on the exact reduced likelihood, see Ly et al. (2018).
}
