\name{ofc}
\docType{data}
\alias{ofc}
\usage{data(ofc)}
\title{Birth data}
\description{
Birth data for singleton live births with gestational age at least 38 weeks.
}
\format{A data frame with 28829 observations on 8 variables.
  \tabular{lll}{
    \code{Year}  \tab numeric  \tab 2006 -- 2009 \cr
    \code{Sex}   \tab character  \tab 'male' or 'female' \cr
    \code{Gestation}  \tab numeric  \tab Gestational age (in weeks). \cr
    \code{Weight}   \tab numeric  \tab birth weight. \cr
    \code{Length}   \tab numeric  \tab height. \cr
    \code{Head}   \tab numeric  \tab head size. \cr
    \code{Chest}   \tab numeric  \tab chest size. \cr
    \code{Region}   \tab character  \tab region. \cr
  }
}
\references{
Wang, CSDA and JSS papers.
}
\keyword{datasets}
