% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson.R
\name{posteriorRhoStat}
\alias{posteriorRhoStat}
\title{ASDF}
\usage{
posteriorRhoStat(n, r, rho, kappa = 1, alternative = "two.sided")
}
\arguments{
\item{n}{numeric > 0, number of samples}

\item{r}{numeric in (-1, 1), the observed Pearson correlation r in the sample}

\item{kappa}{numeric > 0, the scale of the beta distribution, i.e., beta(1/kappa, 1/kappa)}
}
\value{
Returns the posterior density
}
\description{
ASDF
}
