% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson.R
\name{priorRho}
\alias{priorRho}
\title{Prior for Pearson's rho}
\usage{
priorRho(rho, kappa = 1, alternative = "two.sided", betaA = NULL, betaB = NULL)
}
\arguments{
\item{rho}{numeric in (-1, 1) at which the prior needs to be evaluated}

\item{kappa}{numeric > 0 which provides the scale of}

\item{alternative}{a character string specifying the alternative hypothesis must be one of "two.sided" (default),
"greater" or "less"}
}
\value{
numeric, the density value at rho
}
\description{
Prior for Pearson's rho
}
\examples{
priorRho(0.4)

rhoDomain <- seq(-0.99, 0.99, by=0.01)

# kappa <- 1 refers to the uniform distribution
priorRho(rhoDomain)

# kappa <- 2, e.g., beta(1/2, 1/2), is needed for information consistency
yLine <- priorRho(rhoDomain, kappa=2)

graphics::plot(rhoDomain, yLine, type="l")

# kappa <- 1 refers to the uniform distribution, but "greater" truncates it
yLine <- priorRho(rhoDomain, alternative="greater")

graphics::plot(rhoDomain, yLine, type="l")
}
