% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson.R
\name{pValueFromCor}
\alias{pValueFromCor}
\title{Computes the the p value based on the observed correlation.}
\usage{
pValueFromCor(n, stat, method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{n}{integer representing sample size}

\item{method}{character, "pearson", "kendall" or "spearman}

\item{r}{numeric in (-1, 1) representing the sample correlation}
}
\value{
list of three p-values
}
\description{
Computes the the p value based on the observed correlation.
}
