% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson.R
\name{pValueFromT}
\alias{pValueFromT}
\title{Function returns the p value from correlation.}
\usage{
pValueFromT(t, n1, n2 = 0, var.equal = TRUE)
}
\arguments{
\item{t}{numeric representing observed t-statistic}

\item{n1}{integer representing sample size of the first sample}

\item{n2}{integer representing sample size of the second sample, n2=0 implies that it's a one-sample}

\item{var.equal}{logical, TRUE is equal, cannot have unequal, because we don't have access to s1 and s2}
}
\value{
list of three p-values
}
\description{
Function returns the p value from correlation.
}
