% This is a template R document file
% Distributed under GPL 3 or later

\name{bstats}
\alias{bstats}
\title{R package: bstats}
\description{
In this paackage, some R functions are written for the convenience of class uses.  Especially for my st 315, st 210, st 335, st 475/575}

\author{
  B. Wang  \email{bwang@jaguar1.usouthal.edu}
}
\keyword{stats}

