\name{river}
\docType{data}
\alias{river}
\usage{data(river)}
\title{New York river data}
\description{
This is a data set selected from book "Regression by examples" by Samprit Chatterjee and Ali S. Hadi.
}
\format{In a 1976 study exploring the relationship between water quality and land use, Haith (1976) obtained the measurements on 20 river basins in New York State.  A question of interest here is how the land use around a river basin contributes to the water pollution as measured by the mean nitrogen concentration (mg/liter).
  \tabular{lll}{
    \code{River}  \tab character  \tab  River names\cr
    \code{Agr}   \tab numeric  \tab percentage of land area currently in agricultural use \cr
    \code{Forest}   \tab numeric  \tab percentage of forest land \cr
    \code{Rsdntial}   \tab numeric  \tab percentage of land area in residential use \cr
    \code{ComIndl}   \tab numeric  \tab percentage of land area either in commercial or industrial use \cr
    \code{Nitrogen}   \tab numeric  \tab  mean nitrogen concentration\cr
  }
}
\references{
"Regression analysis by example" by Samprit Chatterjee
and Ali S. Hadi, Wiley. ISBN: 978-0-471-74696-6.
}
\keyword{datasets}
