\name{supervisor}
\docType{data}
\alias{supervisor}
\usage{data(supervisor)}
\title{Supervisor performance data}
\description{
This is a data set selected from book "Regression by examples" by Samprit Chatterjee and Ali S. Hadi.
}
\format{A data frame with 28829 observations on 8 variables.
  \tabular{lll}{
    \code{Y}  \tab numeric  \tab overall rating of jon being done by supervisor \cr
    \code{X1--X6}   \tab numeric  \tab average score for six different aspects \cr
  }
}
\references{
"Regression analysis by example" by Samprit Chatterjee
and Ali S. Hadi, Wiley. ISBN: 978-0-471-74696-6.
}
\keyword{datasets}
